[Refactor] TownSettingView(동네설정뷰) 분리
기능별 View 분리
- 동네 선택 버튼
- 거리 설정 슬라이더
