+++ BookBridge/BookBridge/BookBridgeApp.swift
@@ -31,7 +31,7 @@ struct BookBridgeApp: App {
     
     var body: some Scene {
         WindowGroup {
-            HomeView()
+            TownSettingView()
                 .onOpenURL { url in // 뷰가 속한 Window에 대한 URL을 받았을 때 호출할 Handler를 등록하는 함수
                     if AuthApi.isKakaoTalkLoginUrl(url) {
                         _ = AuthController.handleOpenUrl(url: url)
